+++ src/features/file/api/fileDownloadApi.ts
@@ -1,0 +1,7 @@
+import api from '@/shared/api/axiosInstance';
+import type { FileDownloadUrlResponse } from '../types/fileApiTypes';
+
+export const fetchFileDownloadUrl = async (fileId: string): Promise<FileDownloadUrlResponse> => {
+  const { data } = await api.get(`/files/${fileId}/download-url`);
+  return data;
+};
