+++ src/features/file/types/fileApiTypes.ts
@@ -1,0 +1,8 @@
+export interface FileDownloadUrlResponse {
+  fileId: string;
+  key: string;
+  url: string;
+  method: 'PUT' | 'GET';
+  headers: Record<string, string>;
+  expiresInSeconds: number;
+}
