+++ src/features/task-detail/api/fileUploadApi.ts
@@ -2,7 +2,6 @@ import api from '@/shared/api/axiosInstance';
 import type {
   CompleteFileUploadRequest,
   CompleteFileUploadResponse,
-  FileDownloadUrlResponse,
   FileUploadUrlRequest,
   FileUploadUrlResponse,
 } from '@/features/task-detail/types/fileApiTypes';
@@ -31,9 +30,5 @@ export const fileUploadApi = {
       sizeBytes,
     });
     return res.data;
-  },
-  fetchFileDownloadUrl: async (fileId: string): Promise<FileDownloadUrlResponse> => {
-    const { data } = await api.get(`/files/${fileId}/download-url`);
-    return data;
   },
 };
