+++ src/features/task-detail/types/fileApiTypes.ts
@@ -24,12 +24,3 @@ export interface CompleteFileUploadResponse {
   status: string;
   completedAt: string;
 }
-
-export interface FileDownloadUrlResponse {
-  fileId: string;
-  key: string;
-  url: string;
-  method: 'PUT' | 'GET';
-  headers: Record<string, string>;
-  expiresInSeconds: number;
-}
